[EventDispatcher] separation of eventDispatcher from IProvider and updating references
