+++ src/Providers.ts
@@ -5,7 +5,8 @@
  * -------------------------------------------------------------------------------------------
  */
 
-import { EventDispatcher, EventHandler, IProvider } from './providers/IProvider';
+import { EventDispatcher } from './providers/EventDispatcher';
+import { EventHandler, IProvider } from './providers/IProvider';
 /**
  * Provides implementation for acquiring the necessary access token for calling the Microsoft Graph APIs.
  *
